+++ leeseo/src/main/java/com/example/leeseo/domain/mission/dto/MemberMissionResDTO.java
@@ -1,9 +1,12 @@
 package com.example.leeseo.domain.mission.dto;
 
+import com.example.leeseo.domain.mission.enums.MissionStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.List;
 
 public class MemberMissionResDTO {
 
@@ -12,6 +15,28 @@ public class MemberMissionResDTO {
     public record JoinDTO(
         Long member_mission_id,
         LocalDateTime createdAt
+    ){
+    }
+
+    @Builder
+    public record MyMissionListDTO(
+            List<MyMissionDTO> myMissionList,
+            Integer listSize,
+            Integer totalPage,
+            Long totalElements,
+            Boolean isFirst,
+            Boolean isLast
+    ){
+    }
+
+    @Builder
+    public record MyMissionDTO(
+            long memberMissionId,
+            MissionStatus status,
+            LocalDate deadline,
+            String conditional,
+            Integer point,
+            LocalDateTime createdAt
     ){
     }
 }
